+++ MrPiattoClient/Models/Restaurant.cs
@@ -15,12 +15,12 @@ namespace MrPiattoClient.Models
     public class Restaurant
     {
         public int IDRestaurant { get; set; }
-        public int Rating { get; set; }
+        public double Rating { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
         public string Categories { get; set; }
 
-        public Restaurant(int rating, string name, string address, string categorie)
+        public Restaurant(double rating, string name, string address, string categorie)
         {
             Rating = rating;
             Name = name;
